src/requirements.txt
dj-database-url==0.5.0
Django==2.2.7
django-crispy-forms==1.9.0
django-heroku==0.3.1
django-storages==1.9.1
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.1
urllib3==1.25.8
whitenoise==5.0.1
